+++ src/components/PinField/tests/PinField.stories.tsx
@@ -1,6 +1,5 @@
 import { Example } from "utilities/storybook";
 import PinField from "components/PinField";
-import FormControl from "components/FormControl";
 
 export default {
 	title: "Components/PinField",
@@ -15,27 +14,42 @@ export default {
 export const base = () => (
 	<Example>
 		<Example.Item title="no value">
-			<PinField name="pin" />
+			<PinField name="pin" inputAttributes={{ "aria-label": "Pin" }} />
 		</Example.Item>
 
 		<Example.Item title="defaultValue: 1234">
-			<PinField name="pin2" defaultValue="1234" />
+			<PinField name="pin2" defaultValue="1234" inputAttributes={{ "aria-label": "Pin" }} />
 		</Example.Item>
 
 		<Example.Item title="value: 12">
-			<PinField name="pin3" value="12" />
+			<PinField name="pin3" value="12" inputAttributes={{ "aria-label": "Pin" }} />
 		</Example.Item>
 
 		<Example.Item title="defaultValue: 12, valueLength: 6">
-			<PinField name="pin4" defaultValue="12" valueLength={6} />
+			<PinField
+				name="pin4"
+				defaultValue="12"
+				valueLength={6}
+				inputAttributes={{ "aria-label": "Pin" }}
+			/>
 		</Example.Item>
 
 		<Example.Item title="defaultValue: ab, charPattern: alphabetic">
-			<PinField name="pin5" defaultValue="ab" pattern="alphabetic" />
+			<PinField
+				name="pin5"
+				defaultValue="ab"
+				pattern="alphabetic"
+				inputAttributes={{ "aria-label": "Pin" }}
+			/>
 		</Example.Item>
 
 		<Example.Item title="defaultValue: ab, charPattern: alphanumeric">
-			<PinField name="pin6" defaultValue="ab" pattern="alphanumeric" />
+			<PinField
+				name="pin6"
+				defaultValue="ab"
+				pattern="alphanumeric"
+				inputAttributes={{ "aria-label": "Pin" }}
+			/>
 		</Example.Item>
 	</Example>
 );
@@ -43,7 +57,7 @@ export const base = () => (
 export const variant = () => (
 	<Example>
 		<Example.Item title="variant: faded">
-			<PinField name="pin" variant="faded" />
+			<PinField name="pin" variant="faded" inputAttributes={{ "aria-label": "Pin" }} />
 		</Example.Item>
 	</Example>
 );
@@ -51,26 +65,19 @@ export const variant = () => (
 export const size = () => (
 	<Example>
 		<Example.Item title="size: large">
-			<PinField name="pin" size="large" />
+			<PinField name="pin" size="large" inputAttributes={{ "aria-label": "Pin" }} />
 		</Example.Item>
 
 		<Example.Item title="size: xlarge">
-			<PinField name="pin" size="xlarge" />
+			<PinField name="pin" size="xlarge" inputAttributes={{ "aria-label": "Pin" }} />
 		</Example.Item>
 
 		<Example.Item title="size: responsive, s: medium, m+: xlarge">
-			<PinField name="pin" size={{ s: "medium", m: "xlarge" }} />
-		</Example.Item>
-	</Example>
-);
-
-export const formControl = () => (
-	<Example>
-		<Example.Item title="with form control">
-			<FormControl>
-				<FormControl.Label>Label</FormControl.Label>
-				<PinField name="pin" />
-			</FormControl>
+			<PinField
+				name="pin"
+				size={{ s: "medium", m: "xlarge" }}
+				inputAttributes={{ "aria-label": "Pin" }}
+			/>
 		</Example.Item>
 	</Example>
 );