+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - ContextMenu: Added onOpen/onClose handlers support
 - lockScroll + ContextMenu: fixed lockScroll scrollable parent detection
 - Loader: Added ariaLabel prop
+- Progress: Added ariaLabel prop